// types/web3Typings/SpinPurchaser.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import type BN from "bn.js";
import type { ContractOptions } from "web3-eth-contract";
import type { EventLog } from "web3-core";
import type { EventEmitter } from "events";
import type {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export type OwnershipTransferred = ContractEventLog<{
  previousOwner: string;
  newOwner: string;
  0: string;
  1: string;
}>;
export type SpinPurchased = ContractEventLog<{
  amount: string;
  purchaser: string;
  0: string;
  1: string;
}>;

export interface SpinPurchaser extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): SpinPurchaser;
  clone(): SpinPurchaser;
  methods: {
    getLatestPrice(): NonPayableTransactionObject<string>;

    nftPass(): NonPayableTransactionObject<string>;

    owner(): NonPayableTransactionObject<string>;

    passHolderCostPerSpin(): NonPayableTransactionObject<string>;

    purchaseSpins(
      _amount: number | string | BN,
      _value: number | string | BN
    ): NonPayableTransactionObject<void>;

    renounceOwnership(): NonPayableTransactionObject<void>;

    spinsPurchased(arg0: string): NonPayableTransactionObject<string>;

    standardCostPerSpin(): NonPayableTransactionObject<string>;

    transferOwnership(newOwner: string): NonPayableTransactionObject<void>;

    withdraw(): NonPayableTransactionObject<void>;
  };
  events: {
    OwnershipTransferred(cb?: Callback<OwnershipTransferred>): EventEmitter;
    OwnershipTransferred(
      options?: EventOptions,
      cb?: Callback<OwnershipTransferred>
    ): EventEmitter;

    SpinPurchased(cb?: Callback<SpinPurchased>): EventEmitter;
    SpinPurchased(
      options?: EventOptions,
      cb?: Callback<SpinPurchased>
    ): EventEmitter;

    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };

  once(event: "OwnershipTransferred", cb: Callback<OwnershipTransferred>): void;
  once(
    event: "OwnershipTransferred",
    options: EventOptions,
    cb: Callback<OwnershipTransferred>
  ): void;

  once(event: "SpinPurchased", cb: Callback<SpinPurchased>): void;
  once(
    event: "SpinPurchased",
    options: EventOptions,
    cb: Callback<SpinPurchased>
  ): void;
}